requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohttp==3.8.6
aiosignal==1.3.1
albumentations==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.2.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1702571698061/work
black==21.4b2
brotlipy==0.7.0
cachetools==5.3.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1700303426725/work/certifi
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
cloths-segmentation==0.0.2
cloudpickle==2.2.1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
cryptography @ file:///C:/b/abs_8ecplyc3n2/croot/cryptography_1677533105000/work
cycler==0.11.0
Cython==3.0.8
debugpy @ file:///D:/bld/debugpy_1660619187469/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
-e git+https://github.com/facebookresearch/detectron2.git@3ff5dd1cff4417af07097064813c9f28d7461d3c#egg=detectron2
dill==0.2.7.1
dlib==19.24.2
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
exif==1.6.0
filelock==3.12.2
fire==0.5.0
Flask==2.2.5
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
fvcore==0.1.5.post20221221
gast==0.2.2
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.26.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==2.10.0
huggingface-hub==0.16.4
hydra-core==1.3.2
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
iglovikov-helper-functions==0.0.53
imagecorruptions==1.1.2
imageio==2.31.2
importlib-metadata==6.7.0
importlib-resources==5.12.0
imutils==0.5.4
iopath==0.1.9
ipykernel @ file:///D:/bld/ipykernel_1666723449132/work
ipython @ file:///D:/bld/ipython_1651240762678/work
itsdangerous==2.1.2
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.3
joblib==1.3.2
jpeg4py==0.1.4
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1673615989977/work
jupyter_core @ file:///D:/bld/jupyter_core_1658332454038/work
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lightning-utilities==0.10.1
llvmlite==0.39.1
Markdown==3.4.4
MarkupSafe==2.1.4
matplotlib==3.5.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186163140/work
mkl-service==2.4.0
mock==5.1.0
msgpack==1.0.5
msgpack-numpy==0.4.8
multidict==6.0.4
munch==4.0.0
mypy-extensions==1.0.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705352640985/work
networkx==2.6.3
numba==0.56.4
numpy==1.18.5
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1696202382185/work
pandas==1.3.5
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pathspec==0.11.2
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.5.0
platformdirs==4.0.0
plum-py==0.8.7
plumbum==1.8.2
portalocker==2.7.0
pretrainedmodels==0.7.4
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1702399386289/work
protobuf==3.20.3
psutil==5.9.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydot==2.0.0
pyemd==1.0.0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1700607939962/work
pyOpenSSL @ file:///C:/b/abs_552w85x1jz/croot/pyopenssl_1677607703691/work
pyparsing==3.1.1
pyreadline==2.1
PySocks @ file:///C:/ci/pysocks_1594394709107/work
pytesseract==0.3.10
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytorch-lightning==1.9.5
pytorch-toolbelt==0.6.3
pytz==2023.3.post1
PyWavelets==1.3.0
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
qudida==0.0.4
regex==2023.12.25
requests @ file:///C:/ci/requests_1657735288441/work
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
segmentation-models-pytorch==0.3.3
sentry-sdk==1.39.2
setproctitle==1.3.3
shapely==2.0.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
slidingwindow==0.0.14
smmap==5.0.1
tabulate==0.9.0
tensorboard==1.14.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorpack==0.11
termcolor==2.3.0
tf-pose==0.11.0
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.9.2
tk==0.1.0
toml==0.10.2
tomli==2.0.1
torch==1.13.1
torchaudio==0.13.1
torchmetrics==0.11.4
torchvision==0.14.1
tornado @ file:///D:/bld/tornado_1656937938087/work
tqdm==4.66.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
typed-ast==1.5.5
typing_extensions==4.7.1
urllib3 @ file:///C:/b/abs_9bcwxczrvm/croot/urllib3_1673575521331/work
wandb==0.16.2
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1699959196938/work
Werkzeug==2.2.3
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306165655/work
wincertstore==0.2
wrapt==1.16.0
yacs==0.1.8
yarl==1.9.4
zipp==3.15.0